// src/app/interfaces/models.ts
// Generated by https://quicktype.io

export interface Voyage {
    isActive?:     boolean;
    price:        number;
    picture:      string;
    passagers:    Passager[];
    depart:       string;
    arrivee:      string;
    company:      string;
    email:        string;
    phone:        string;
    address:      string;
    about:        string;
    registered?:   string;
    heureDepart:  string;
    heureArrivee: string;
    nbrePlaces:   number;
    takePlaces?:   string;
    numeroBus: string;
    userId: string;
    $key?: string;
}

export interface Passager {
    passagerId?:   string;
    name:         string;
    prenom:       string;
    paid:         number;
    place:        Number[];
    modePaiement?: string;
    qte:          number;
    codeMoney?:    string;
}

export interface User {
    $key?: string;
    email: string;
    code: string;
    uuid: string;
}